update kbsig to also print counter, add comments on why counter is not updating correctly
Crux of the problem is the fact that the read() call in the signal handler is
blocking since stdin is emitting SIGPOLL when it is ready, not when it actually
has data to read. this means majority of the time we are blocked in the signal
handler and the counter is only incremented after we read input.

This is why poll() + pthreads is a much better solution...